Tidy up into a service
